backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
psycopg2-binary==2.9.10
bcrypt==4.0.1
python-dotenv==1.0.0